Share buffers using Prime handles.
We currently share buffers using names/flinks. Instead of this we
should be really sharing buffers using prime handles/dma_buf fds.
This patch makes changes to support prime handles. Also, note that
creating buffers with name is still supported in case ENABLE_FLINK_SUPPORT
is set to 1 in the board configuration(Support for Prime handle is
disabled in this case).